requirements.txt
featuretools[dask]
feature-engine
ipython
lightgbm
numpy
pandas
seaborn
matplotlib
scikit-learn
sklearn-genetic-opt
tsfresh
xgboost
woodwork
boruta
optuna
plotly
catboost
shap
